# api/openapi.yml
openapi: 3.0.3
info:
  title: Manjo Ticketing API
  version: 1.0.0
  description: |
    Complete HTTP surface of the Go API (`backend/`).

    ## The envelope

    Every JSON endpoint answers `{ "code": 200000, "message": "Success", "data": ... }` on success
    and `{ "code": <error code>, "message": ..., "data": ... }` on failure — the HTTP status line
    and the envelope code both carry meaning, and a client that branches only on status will miss
    distinctions the numeric code makes.

    Two endpoints sit outside it: `GET /healthz` (bare JSON, read by orchestrators) and
    `GET /api/v1/ticket/order/{order_id}/qris.png` (a PNG).

    ## Identifiers

    Public routes address an **event by its slug** and an **order by its order number**. Admin
    routes address everything by **UUID**. The two namespaces are not interchangeable — an order
    has both a public order number and an admin-side UUID, and passing one where the other is
    expected is a 400 or a 404, never a silent success.

    ## Authentication

    `/api/v1/admin/*` requires a bearer JWT from `POST /api/v1/admin/login`. `POST /v1.0/callback/exec`
    is the payment gateway's own endpoint and authenticates with a *different* bearer credential
    (`PG_CALLBACK_TOKEN`). Everything else is deliberately unauthenticated — the guest purchase
    journey has no accounts (Constitution Principle VI).

    ## Rate limits

...
    per-IP budget, and the SSE status stream caps concurrent connections per IP rather than
...

...
servers:
...
security: []
...
paths:
...
  /api/v1/admin/orders:
    get:
      tags: [Admin - Orders & Attendees]
      summary: List orders
      operationId: adminListOrders
      security: [{ adminAuth: [] }]
      description: |
        Both filters are optional and combine, and both are applied BEFORE paging — so `total`
        counts what the filter matches, not what the table holds.

        Note the deliberate asymmetry with the paging parameters: a malformed `status` or
        `event_id` is a 400, because the caller named something that does not exist, while a
        malformed or out-of-range `page` is corrected, because the caller named a position that
        drifted.
      parameters:
        - name: status
          in: query
          required: false
          schema:
            type: string
            enum: [PENDING, PAID, CANCELLED, EXPIRED]
        - name: event_id
          in: query
          required: false
          description: |
            Resolved to the event's ticket types before matching. An event with no ticket types
            yields `total: 0`, not an unfiltered count.
          schema:
            type: string
            format: uuid
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
      responses:
        '200':
          description: One page of matching orders, newest first.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - type: object
                    properties:
                      data:
                        allOf:
                          - $ref: '#/components/schemas/PageMeta'
                          - type: object
                            required: [items]
                            properties:
                              items:
                                type: array
                                items:
                                  $ref: '#/components/schemas/OrderSummary'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/v1/admin/attendees:
    get:
      tags: [Admin - Orders & Attendees]
      summary: List attendees
      operationId: adminListAttendees
      security: [{ adminAuth: [] }]
      description: |
        Ticket holders with their submitted details. Both filters are optional, and both are
        applied before paging.
      parameters:
        - name: order_id
          in: query
          required: false
          description: Order **UUID**, not the order number.
          schema:
            type: string
            format: uuid
        - name: event_id
          in: query
          required: false
          schema:
            type: string
            format: uuid
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
      responses:
        '200':
          description: One page of matching attendees.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - type: object
                    properties:
                      data:
                        allOf:
                          - $ref: '#/components/schemas/PageMeta'
                          - type: object
                            required: [items]
                            properties:
                              items:
                                type: array
                                items:
                                  $ref: '#/components/schemas/AttendeeSummary'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /api/v1/admin/fees:
    get:
      tags: [Admin - Fees]
      summary: List fees
      operationId: adminListFees
      security: [{ adminAuth: [] }]
      description: One page of the fee master rows, in display order.
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
      responses:
        '200':
          description: One page of fees.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - type: object
                    properties:
                      data:
                        allOf:
                          - $ref: '#/components/schemas/PageMeta'
                          - type: object
                            required: [items]
                            properties:
                              items:
                                type: array
                                items:
                                  $ref: '#/components/schemas/FeeAdminView'
        '401':
          $ref: '#/components/responses/Unauthorized'
    post:
      tags: [Admin - Fees]
      summary: Create a fee
      operationId: adminCreateFee
      security: [{ adminAuth: [] }]
      description: |
        Applies to orders booked from now on. A fee is frozen onto an order at booking time — with
        any percentage baked into the name, e.g. "PPN (11%)" — so editing one never changes an
        existing order's total.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FeeRequest'
      responses:
        '201':
          description: Created.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/FeeAdminView'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/v1/admin/fees/{id}:
    parameters:
      - $ref: '#/components/parameters/ResourceId'
    put:
      tags: [Admin - Fees]
      summary: Update a fee
      operationId: adminUpdateFee
      security: [{ adminAuth: [] }]
      description: Full replacement. Existing orders keep the fee lines they were sold with.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FeeRequest'
      responses:
        '200':
          description: Updated.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/FeeAdminView'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags: [Admin - Fees]
      summary: Delete a fee
      operationId: adminDeleteFee
      security: [{ adminAuth: [] }]
      responses:
        '200':
          description: Deleted.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          message:
                            type: string
                            example: Fee deleted.
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    adminAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        The token from `POST /api/v1/admin/login`. HS256, verified against `JWT_SECRET`; only
        HS256 is accepted, because letting the token's own header select the algorithm is what
        makes algorithm-confusion forgeries possible.
    gatewayCallbackAuth:
      type: http
      scheme: bearer
      description: |
        `PG_CALLBACK_TOKEN` — the payment gateway's credential, entirely separate from the admin
        JWT. Compared in constant time, and unconditionally: returning early on the first differing
        byte would leak the expected token a character at a time to anyone able to time the
        endpoint.
  parameters:
...
    ResourceId:
      name: id
      in: path
      required: true
      description: The resource's UUID.
      schema:
        type: string
        format: uuid
...
    Page:
      name: page
      in: query
      required: false
      description: |
        1-based page number. Corrected rather than refused: zero, negative, non-numeric and
        absent all resolve to page 1, and a page past the last resolves to the last page. Read
        the returned `page` to see which one was actually served — a stale bookmark self-corrects
        instead of erroring.
      schema:
        type: integer
        minimum: 1
        default: 1
    PageSize:
      name: page_size
      in: query
      required: false
      description: |
        Rows per page. Values above the maximum are reduced to it; values below 1, non-numeric
        values and absent all take the default. Never refused. The cap is enforced server-side:
        no request can demand an unbounded page.
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
...
  responses:
    ValidationError:
      description: |
        The request was malformed. `data` carries a field map when more than one field is at fault.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            code: 400001
            message: Some fields are missing or invalid.
            data:
              attendees[0].phone: Phone number must be 12-15 digits.
...
    Unauthorized:
      description: No bearer token, or the token is invalid or expired.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            code: 401001
            message: A bearer token is required.
            data: null
    NotFound:
      description: No such resource.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            code: 404001
            message: Event not found.
            data: null
...
  schemas:
    PageMeta:
      type: object
      description: |
        Where a page sits in its list. Carried alongside `items` in `data` on every paginated
        admin list. `page` and `page_size` are what the server SERVED, which may differ from
        what was requested — see the `page` parameter.
      required: [page, page_size, total, total_pages]
      properties:
        page:
          type: integer
          minimum: 1
          example: 2
        page_size:
          type: integer
          minimum: 1
          maximum: 100
          example: 20
        total:
          type: integer
          format: int64
          minimum: 0
          description: Rows matching the current filters, ignoring paging.
          example: 137
        total_pages:
          type: integer
          minimum: 0
          description: Zero when `total` is zero — never "page 1 of 0".
          example: 7
...
    Envelope:
      type: object
      description: |
        The one JSON shape every endpoint returns. `code` is 200000 on success regardless of which
        2xx the status line carries.
      required: [code, message, data]
      properties:
        code:
          type: integer
          example: 200000
        message:
          type: string
          example: Success
        data:
          nullable: true
          description: The typed payload, or null.
    Error:
      type: object
      description: |
        The failure envelope. `data` is null except where a contract attaches detail — a validation
        field map, the current QR payload on PAYMENT_ALREADY_STARTED, the remaining cooldown on a
        guest resend refusal.
      required: [code, message, data]
      properties:
        code:
          type: integer
          description: |
            The numeric envelope code. Note that several distinct conditions render as 400001, so
            a client that must tell them apart reads the stable string code where one is offered
            (see `AvailabilityReason.code`).
          example: 400001
        message:
          type: string
        data:
          nullable: true
    Money:
      type: string
      description: |
        A decimal amount as a string, so no precision is lost in transit. Responses are always
        strings; requests also accept a JSON number.
      example: "150000.00"

    # --- Public catalogue ---------------------------------------------------
...
    OrderSummary:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: The admin-side UUID — what the resend and payment-ops routes take.
        order_number:
          type: string
          description: The public identifier the guest sees.
        buyer_name: { type: string }
        buyer_email: { type: string }
        status: { type: string, enum: [PENDING, PAID, CANCELLED, EXPIRED] }
        total_amount: { $ref: '#/components/schemas/Money' }
        created_at: { type: string, format: date-time, nullable: true }
    AttendeeSummary:
      type: object
      properties:
        name: { type: string }
        email: { type: string }
        ticket_type_name: { type: string }
        order_number: { type: string }
    FeeRequest:
      type: object
      required: [name, fee_type, value, position, is_active]
      properties:
        name: { type: string, example: PPN }
        fee_type: { type: string, enum: [PERCENT, FIXED] }
        value:
          allOf: [{ $ref: '#/components/schemas/Money' }]
          description: A percentage for PERCENT, an amount for FIXED. Never negative.
          example: "11"
        position:
          type: integer
          format: int32
          description: Orders the lines on the order summary.
        is_active: { type: boolean }
    FeeAdminView:
      type: object
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        fee_type: { type: string, enum: [PERCENT, FIXED] }
        value: { $ref: '#/components/schemas/Money' }
        position: { type: integer, format: int32 }
        is_active: { type: boolean }
        created_at: { type: string, format: date-time, nullable: true }
        updated_at: { type: string, format: date-time, nullable: true }
...